Move request handler into its own files
